+++ src/components/atoms/issueMessage/issueMessage.scss
@@ -1,5 +1,5 @@
 .issueMessage {
-	@include description;
-	color: $secondary-utility-color;
-	margin-top: $x-small-space;
+  @include description;
+  color: $secondary-utility-color;
+  margin-top: $x-small-space;
 }
