+++ src/components/atoms/localisationIncidents/LocalisationIncidents.jsx
@@ -1,8 +1,7 @@
 import React from 'react'
 
 const LocalisationIncidents = ({cssClass, text }) => {
-
-return <span className={`localisation-component ${cssClass}`}>{text}</span>
+  return <span className={`localisation-component ${cssClass}`}>{text}</span>
 }
 
 export default LocalisationIncidents;
