+++ src/components/atoms/tagsIncidents/TagsIncidents.jsx
@@ -3,10 +3,10 @@ import React from 'react'
 const TagsIncidents = ({cssClass, text }) => {
 
 return(
-    <div className="tag-container">
-        <span className={`tags-component ${cssClass}`}>{text}</span> 
-    </div>
-    )
+  <div className="tag-container">
+    <span className={`tags-component ${cssClass}`}>{text}</span> 
+  </div>
+  )
 }
 
 export default TagsIncidents;
