+++ src/components/atoms/title/title.jsx
@@ -2,9 +2,9 @@ import React from "react";
 
 const Title = ({children, cssClass}) => {
 
-	return (
-		<p className={`title-component ${cssClass}`}>{children}</p>
-	)
+  return (
+    <p className={`title-component ${cssClass}`}>{children}</p>
+  )
 }
 
 export default Title;
