+++ src/components/molecules/establishmentContact/establishmentManagerContact.jsx
@@ -2,22 +2,22 @@ import React from 'react';
 import Title from '../../atoms/title/title';
 
 const EstablishmentManagerContact = ({firstName, lastName, mail, phone, nbSensors}) => {
-	return (
-		<div className="establishmentContact">
-			<Title cssClass="detailed-information-title">
-				Représentant de l’établissement
-			</Title>
-			<div className="commonData">
-				<p className="name">{firstName} {lastName}</p>
-				<p className="contact">{mail}</p>
-				<p className="contact">{phone}</p>
-			</div>
-			<div className="importantData">
-				<p>{nbSensors}</p>
-				<p>capteurs</p>
-			</div>
-		</div>
-	)
+  return (
+    <div className="establishmentContact">
+      <Title cssClass="detailed-information-title">
+        Représentant de l’établissement
+      </Title>
+      <div className="commonData">
+        <p className="name">{firstName} {lastName}</p>
+        <p className="contact">{mail}</p>
+        <p className="contact">{phone}</p>
+      </div>
+      <div className="importantData">
+        <p>{nbSensors}</p>
+        <p>capteurs</p>
+      </div>
+    </div>
+  )
 }
 
 export default EstablishmentManagerContact;
