+++ src/components/pages/Agenda.jsx
@@ -2,17 +2,18 @@ import React, { useState } from 'react'
 import CalendarList from '../organisms/calendarList/CalendarList'
 import NewInterventionForm from '../organisms/newInterventionForm/NewInterventionForm'
 import Modal from '../templates/modal/Modal'
+import Page from '../templates/page/Page'
 
 const Agenda = () => {
   const [showModal, setShowModal] = useState(true)
 
   return (
-    <div className="agenda">
-    <CalendarList/>
-    <Modal showModal={showModal}>
-      <NewInterventionForm cancelCallback={setShowModal} validateCallback={setShowModal}/>
-    </Modal>
-    </div>
+    <Page>
+      <CalendarList/>
+      <Modal showModal={showModal}>
+        <NewInterventionForm cancelCallback={setShowModal} validateCallback={setShowModal}/>
+      </Modal>
+    </Page>
   );
 }
