+++ src/components/pages/Routes.jsx
@@ -1,18 +1,18 @@
-import React, { Fragment } from 'react';
+import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import Sidebar from '../organisms/sidebar/Sidebar';
-import Home from './Home';
+import Home from './home/Home';
 import Agenda from './Agenda';
 
 const Routes = () => {
   return (
-    <Fragment>
+    <>
       <Sidebar/>
       <Switch>
         <Route path="/" exact component={Home} />
         <Route path="/agenda" exact component={Agenda} />
       </Switch>
-    </Fragment>
+    </>
   );
 }
