+++ src/components/templates/page/Page.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Page = ({children, className}) => {
+  return (
+    <div className={`page${className ? ' ' + className : ''}`}>
+      {children}
+    </div>
+  );
+}
+
+export default Page;
